# api/openapi.yaml
openapi: 3.1.0
info:
  title: EZ VPN Lego Control API
  version: 0.1.0-draft
  description: Local control-plane contract used by official adapters.
servers:
  - url: http://localhost/v1
    description: HTTP transported over the local Unix socket
paths:
  /health:
    get:
      operationId: getHealth
      responses:
        "200":
          description: Control-plane health
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Health"
  /bootstrap/owner:
    post:
      operationId: bootstrapOwner
      description: Creates the only owner when no administrator exists.
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Owner created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
        "409":
          description: An administrator already exists
  /bootstrap/vpn:
    post:
      operationId: bootstrapVPN
      description: Creates the initial REALITY/Vision instance and owner access.
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/AdminID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BootstrapVPN"
      responses:
        "201":
          description: VPN instance created and validated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessResult"
        "409":
          description: VPN instance already exists
  /access:
    post:
      operationId: createAccess
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/AdminID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "201":
          description: Access created and active Xray configuration reconciled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccessResult"
  /auth/telegram/pairing:
    post:
      operationId: createTelegramPairing
      description: Creates a single-use pairing token valid for 15 minutes.
      security: [{bearerAuth: []}]
      parameters:
        - $ref: "#/components/parameters/AdminID"
      responses:
        "201":
          description: Pairing token created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PairingToken"
  /auth/telegram/claim:
    post:
      operationId: claimTelegramPairing
      security: [{bearerAuth: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token, subject]
              properties:
                token: {type: string}
                subject: {type: string}
      responses:
        "200":
          description: Telegram identity bound to administrator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
  /auth/telegram/resolve:
    get:
      operationId: resolveTelegramIdentity
      security: [{bearerAuth: []}]
      parameters:
        - name: subject
          in: query
          required: true
          schema: {type: string}
      responses:
        "200":
          description: Active administrator identity
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Admin"
  /users:
    get:
      operationId: listUsers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AdminID"
      responses:
        "200":
          description: Users visible to the authenticated administrator
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AdminID"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUser"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCreation"
        "200":
          description: Idempotent replay of an existing result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserCreation"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    AdminID:
      name: X-Admin-ID
      in: header
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
  schemas:
    Admin:
      type: object
      additionalProperties: false
      required: [id, role, status, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, operator, viewer]
        status:
          type: string
          enum: [active, disabled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Health:
      type: object
      additionalProperties: false
      required: [status, version]
      properties:
        status:
          type: string
          enum: [healthy, degraded]
        version:
          type: string
    User:
      type: object
      additionalProperties: false
      required: [id, name, status, created_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [active, disabled]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    CreateUser:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 80
    UserCreation:
      type: object
      additionalProperties: false
      required: [operation_id, user, replayed]
      properties:
        operation_id:
          type: string
          format: uuid
        user:
          $ref: "#/components/schemas/User"
        replayed:
          type: boolean
    BootstrapVPN:
      type: object
      additionalProperties: false
      required: [public_address, port, target, server_name]
      properties:
        public_address: {type: string}
        port: {type: integer, minimum: 1, maximum: 65535}
        target: {type: string}
        server_name: {type: string}
    AccessResult:
      type: object
      additionalProperties: false
      required: [operation_id, user_id, device_id, uri]
      properties:
        operation_id: {type: string, format: uuid}
        user_id: {type: string, format: uuid}
        device_id: {type: string, format: uuid}
        uri: {type: string}
    PairingToken:
      type: object
      additionalProperties: false
      required: [token, expires_at]
      properties:
        token: {type: string}
        expires_at: {type: string, format: date-time}