# internal/dbgen/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package dbgen

import (
	"github.com/jackc/pgx/v5/pgtype"
)

type Admin struct {
	ID        string             `json:"id"`
	Role      string             `json:"role"`
	Status    string             `json:"status"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
}

type AdminIdentity struct {
	ID        string             `json:"id"`
	AdminID   string             `json:"admin_id"`
	Provider  string             `json:"provider"`
	Subject   string             `json:"subject"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type AdminPairingToken struct {
	TokenHash  []byte             `json:"token_hash"`
	AdminID    string             `json:"admin_id"`
	ExpiresAt  pgtype.Timestamptz `json:"expires_at"`
	ConsumedAt pgtype.Timestamptz `json:"consumed_at"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
}

type AuditEvent struct {
	Sequence     int64              `json:"sequence"`
	ID           string             `json:"id"`
	ActorAdminID pgtype.UUID        `json:"actor_admin_id"`
	Action       string             `json:"action"`
	ResourceType string             `json:"resource_type"`
	ResourceID   pgtype.UUID        `json:"resource_id"`
	Outcome      string             `json:"outcome"`
	Metadata     []byte             `json:"metadata"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
}

type Backup struct {
	ID            string             `json:"id"`
	FormatVersion int32              `json:"format_version"`
	Path          string             `json:"path"`
	Checksum      string             `json:"checksum"`
	Status        string             `json:"status"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	VerifiedAt    pgtype.Timestamptz `json:"verified_at"`
}

type ConfigRevision struct {
	ID          string             `json:"id"`
	NodeID      string             `json:"node_id"`
	Sequence    pgtype.Int8        `json:"sequence"`
	State       string             `json:"state"`
	ContentHash string             `json:"content_hash"`
	Content     []byte             `json:"content"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	VerifiedAt  pgtype.Timestamptz `json:"verified_at"`
}

type Credential struct {
	ID            string             `json:"id"`
	DeviceID      string             `json:"device_id"`
	NodeID        string             `json:"node_id"`
	Driver        string             `json:"driver"`
	SchemaVersion int32              `json:"schema_version"`
	Status        string             `json:"status"`
	Secret        []byte             `json:"secret"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	UpdatedAt     pgtype.Timestamptz `json:"updated_at"`
}

type Device struct {
	ID        string             `json:"id"`
	UserID    string             `json:"user_id"`
	Name      string             `json:"name"`
	Status    string             `json:"status"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	DeletedAt pgtype.Timestamptz `json:"deleted_at"`
}

type Node struct {
	ID            string             `json:"id"`
	Name          string             `json:"name"`
	Status        string             `json:"status"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	UpdatedAt     pgtype.Timestamptz `json:"updated_at"`
	PublicAddress string             `json:"public_address"`
}

type Operation struct {
	ID             string             `json:"id"`
	Kind           string             `json:"kind"`
	Status         string             `json:"status"`
	RequestedBy    pgtype.UUID        `json:"requested_by"`
	IdempotencyKey string             `json:"idempotency_key"`
	Input          []byte             `json:"input"`
	Result         []byte             `json:"result"`
	ErrorCode      pgtype.Text        `json:"error_code"`
	Attempts       int32              `json:"attempts"`
	AvailableAt    pgtype.Timestamptz `json:"available_at"`
	StartedAt      pgtype.Timestamptz `json:"started_at"`
	FinishedAt     pgtype.Timestamptz `json:"finished_at"`
	CreatedAt      pgtype.Timestamptz `json:"created_at"`
	UpdatedAt      pgtype.Timestamptz `json:"updated_at"`
}

type OutboxEvent struct {
	ID            string             `json:"id"`
	Topic         string             `json:"topic"`
	AggregateType string             `json:"aggregate_type"`
	AggregateID   string             `json:"aggregate_id"`
	Payload       []byte             `json:"payload"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	PublishedAt   pgtype.Timestamptz `json:"published_at"`
	Attempts      int32              `json:"attempts"`
	LastError     pgtype.Text        `json:"last_error"`
}

type ProtocolInstance struct {
	ID            string             `json:"id"`
	NodeID        string             `json:"node_id"`
	Driver        string             `json:"driver"`
	SchemaVersion int32              `json:"schema_version"`
	Status        string             `json:"status"`
	Settings      []byte             `json:"settings"`
	CreatedAt     pgtype.Timestamptz `json:"created_at"`
	UpdatedAt     pgtype.Timestamptz `json:"updated_at"`
}

type Release struct {
	Version     string             `json:"version"`
	Channel     string             `json:"channel"`
	Manifest    []byte             `json:"manifest"`
	Status      string             `json:"status"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	InstalledAt pgtype.Timestamptz `json:"installed_at"`
}

type User struct {
	ID        string             `json:"id"`
	Name      string             `json:"name"`
	Status    string             `json:"status"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UpdatedAt pgtype.Timestamptz `json:"updated_at"`
	DeletedAt pgtype.Timestamptz `json:"deleted_at"`
}